delete representable and rename file
