+++ src/lib/constants.rs
@@ -1,9 +1,9 @@
 #![allow(clippy::non_ascii_literal)]
 
-use super::representable::Searchable;
+use super::searchable::Searchable;
 use macros::Searchable;
 
-use super::representable::get_by_repr;
+use super::searchable::get_by_repr;
 
 #[allow(clippy::upper_case_acronyms)]
 #[derive(PartialEq, Clone, Copy, Debug)]
