+++ src/lib/mod.rs
@@ -2,8 +2,8 @@ pub mod constants;
 pub mod errors;
 mod eval;
 pub mod operators;
-mod representable;
 mod rpn;
+mod searchable;
 mod tokenize;
 pub mod tokens;
 pub mod utils;
